Rename things and restructure project
