+++ examples/hello_world_listener.cpp
@@ -18,37 +18,35 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#ifndef HOUSOU__LISTENER_HPP_
-#define HOUSOU__LISTENER_HPP_
+#include <housou/string_listener.hpp>
 
-#include <housou/base_listener.hpp>
+#include <unistd.h>
 
+#include <iostream>
 #include <string>
 
-namespace housou
+int main()
 {
+  housou::StringListener listener(8080);
 
-class Listener : public BaseListener
-{
-public:
-  explicit Listener(int port)
-  : BaseListener(port)
-  {
+  if (!listener.connect()) {
+    std::cerr << "Failed to connect listener on port " <<
+      listener.port << "!" << std::endl;
+
+    return 1;
   }
 
-  std::string receive(int length)
-  {
-    char * buffer = new char[length];
+  while (true) {
+    auto message = listener.receive(64);
 
-    BaseListener::receive(buffer, length);
+    if (message.size() > 0) {
+      std::cout << "Received: " << message << std::endl;
+    }
 
-    std::string message(buffer);
-    delete[] buffer;
-
-    return message;
+    usleep(100 * 1000);
   }
-};
 
-}  // namespace housou
+  listener.disconnect();
 
-#endif  // HOUSOU__LISTENER_HPP_
+  return 0;
+}
